refactor: move IR types to their own file
