+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from './engine';
 export * from './rules';
 export * from './helpers';
+export * from './ir';
 export * from './types';
 export * from './validator';
 export * from './warning';
